% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa.R
\name{get_ncaa}
\alias{get_ncaa}
\title{Logos and colors of NCAA basketball}
\usage{
get_ncaa()
}
\value{
Colors and logos from NCAA Basketball Div-1
}
\description{
NCAA Basketball Div-1 data table with logos and colors
}
\examples{
get_ncaa ()

}
